+++ NEWS.md
@@ -18,13 +18,22 @@
 - Add confirmation if the number of visit levels is too large. Use `options(mmrm.max_visits = )`
   to specify the maximum number of visits allowed in non-interactive mode, or asked in interactive
   sessions.
+- Previously `model.frame` does not work with new given `data`. It now works with new `data`.
+  In addition, it always use "na.omit" regardless of options or arguments in `model.frame`.
+  `subset` argument is always ignored.
 
 ### Bug Fixes
 
 - Previously `mmrm` fit follows the global option `na.action` and if it is set
   other than "na.omit" error will happen. This is now fixed and `NA` values are
   always removed prior to model fitting.
-
+- Previously `model.frame` call on `mmrm` object with transformed terms, e.g.
+  `mmrm(Y ~ log(X) + ar1(VISIT|ID))`, will cause errors. This is now fixed and
+  `model.frame` returns the data correctly.
+- Previously `mmrm` always requires `data` argument. Now fitting `mmrm` can use
+  environment variables, instead of requiring `data` argument. `fit_mmrm` is not affected.
+- Previously `emmeans` does not work if fixed effect does not contain visit variable.
+  This is now fixed.
 # mmrm 0.2.2
 
 ### New Features